refactor OverallMetricChart
